List to 'event' from Rollup instead of 'change' to restart the Electron process (#6)

+++ scripts/dev-runner.js
@@ -78,7 +78,11 @@ async function start() {
 
   startElectron(RENDERER_URL)
 
-  mainWatcher.on('change', () => {
+  mainWatcher.on('event', (event) => {
+    if (event.code !== 'BUNDLE_END') {
+      return
+    }
+
     if (electronProcess && electronProcess.kill) {
       manualRestart = true
       process.kill(electronProcess.pid)
